third commit-fixing the long int problem on report service

+++ src/main/java/com/finalproject/uni_earn/service/ReportService.java
@@ -6,7 +6,8 @@ import com.finalproject.uni_earn.entity.Reports;
 import com.finalproject.uni_earn.entity.enums.ReportState;
 
 public interface ReportService {
-    public Reports submitReport(int reporterId, int reportedUserId, String feedback);
+    public Reports submitReport(long reporterId, long reportedUserId, String feedback);
+
     public ReportDTO resolveReport(Long reportId, ReportState status);
 
     PaginatedReportDTO getallReports(int page , int size);
